swe plane solving poisson problem and velocity; double dot still has errors.

+++ src/lpm_error_impl.hpp
@@ -109,7 +109,7 @@ struct ReduceErrorFtor<V1, V2, 1> {
 
 template <typename V1, typename V2, typename V3>
 void compute_error(const V1 err, const V2 appx, const V3 exact) {
-  static_assert(V1::Rank == V2::Rank and V2::Rank == V3::Rank,
+  static_assert(static_cast<int>(V1::Rank) == static_cast<int>(V2::Rank) and static_cast<int>(V2::Rank) == static_cast<int>(V3::Rank),
                 "view ranks must match");
   LPM_REQUIRE(err.extent(0) == appx.extent(0) and
               appx.extent(0) == exact.extent(0));
